+++ app/src/main/java/in/andres/kandroid/DashOverdueFragment.java
@@ -6,6 +6,7 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.ListFragment;
 import android.text.Html;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -46,6 +47,7 @@ public class DashOverdueFragment extends ListFragment {
     @Override
     public void onListItemClick(ListView l, View v, int position, long id) {
         super.onListItemClick(l, v, position, id);
+        Log.i(Constants.TAG, "Launching TaskDetailActivity from DashOverdueFragment.");
         KanboardTask clickedTask = ((MainActivity)getActivity()).getDashboard().getOverdueTasks().get(position);
         Intent taskIntent = new Intent(getContext(), TaskDetailActivity.class);
         taskIntent.putExtra("task", clickedTask);