+++ app/src/main/java/in/andres/kandroid/ProjectInactiveTasksFragment.java
@@ -4,6 +4,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -45,6 +46,7 @@ public class ProjectInactiveTasksFragment extends Fragment {
             ((ExpandableListView) getView().findViewById(R.id.expandable_list)).setOnChildClickListener(new ExpandableListView.OnChildClickListener() {
                 @Override
                 public boolean onChildClick(ExpandableListView parent, View v, int groupPosition, int childPosition, long id) {
+                    Log.i(Constants.TAG, "Launching TaskDetailActivity from ProjectInactiveTasksFragment.");
                     KanboardProject project = ((MainActivity) getActivity()).getProject();
                     KanboardTask clickedTask = project.getGroupedInactiveTasks().get(project.getSwimlanes().get(groupPosition).getId()).get(childPosition);
                     Intent taskIntent = new Intent(getContext(), TaskDetailActivity.class);