+++ Changelog.md
@@ -8,6 +8,7 @@
 #### fixes
 * Show string when task is not assigned to a user
 * Show string when category is not assigned
+* Login activity is started twice on first run
 
 ## v0.3.1
 #### fixes